Create killSpam.sh
This is less DFIR, and more system administration. It is meant for admins that use SpamAssassin on their mail servers. It looks at all the files in the amavisd virusmail [0-9a-zA-Z] sub-folder (you must be in the sub-folder first to run the script), providing To, From, Subject, and Date and then asks to release or delete.

It doesn't do all the sub-folders because I had 300+ in each one, and didn't want to start and run through everything.