Create http_extractor.py
This is a complete rewrite of lHttpExtractor.py and cyHttpExtractor.py. Now finds all instances of URI in a file, regardless of how many per line, and prints them out. This version is modular. Tried to follow advice from Wyatt Roersma's review of the old code.

Need to do - create setup.py so it can be imported and ran regardless of environment.